// src/app/core/models/user.model.ts
export interface User {
  email: string;
  phone: string;
  password: string;
  firstName: string;
  lastName: string;
  address: string;
  birthDate: string;
  gender: string;
  name: string;
}
export const user = {
  confirm: {
    email: '',
    password: '',
    name: '',
    id: '',
    country_code: '',
    phone_number: ''
  }
};